requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
httpx
GPUtil
gradio
pystray
pillow
python-dotenv
requests
openai